docker_agent/requirements.txt
docker-py >= 1.8
bottle >= 0.12
portpicker >= 1.1
